Move generic code to base class
